web-scraper/requirements.txt
Flask==2.1.1
pymongo==4.3.3
bs4==0.0.1
google==3.0.0
requests==2.28.1
python-dotenv==0.16.0
pillow == 9.3.0
wordcloud == 1.8.2.2
numpy == 1.23.5
pytest==7.2.0
mongomock==4.1.2
pytest-cov==4.0.0
datetime == 4.7
